You can disable spellcheck or documentation when building docs. (#10377)
This cleans up the document building process and replaces it
with breeze-only. The original instructions with
`pip install -e .[doc]` stopped working so there is no
point keeping them.

Extracted from #10368

+++ scripts/ci/in_container/run_docs_build.sh
@@ -26,9 +26,9 @@ trap "${HANDLERS}${HANDLERS:+;}in_container_fix_ownership" EXIT
 sudo rm -rf "${AIRFLOW_SOURCES}/docs/_build/*"
 sudo rm -rf "${AIRFLOW_SOURCES}/docs/_api/*"
 
-sudo -E "${AIRFLOW_SOURCES}/docs/build"
+sudo -E "${AIRFLOW_SOURCES}/docs/build_docs.py" "${@}"
 
-if [[ ${CI} == "true" ]]; then
+if [[ ${CI} == "true" && -d "${AIRFLOW_SOURCES}/docs/_build/html" ]]; then
     rm -rf "/files/documentation"
     cp -r "${AIRFLOW_SOURCES}/docs/_build/html" "/files/documentation"
 fi
